requirements.txt
# Evaluation metrics dependencies
pandas>=1.3.0
numpy>=1.21.0
Levenshtein>=0.20.0
python-Levenshtein>=0.20.0
jiwer>=3.0.0

# OCR dependencies
pytesseract>=0.3.10
pdf2image>=1.16.0
Pillow>=9.0.0
opencv-python>=4.7.0

# Ollama integration
requests>=2.31.0
ollama>=0.3.0
pydantic>=2.0.0

